Disallow transform self-invalidation
This changes a DependencyManager such that a transform is never
allowed to invalidate itself, even if it's invalidation function
returns true when given itself. This is fundamentally necessary as
"invalidate everything except myself" is not possible to express
abstractly due to the instability of pattern matching
this.getClass (or just limitations of my Scala skills).

Fundamentally, if a transform can invalidate itself then it can never
be scheduled. This updates the invalidation Scaladoc to indicate that
the behavior of that function is only defined (and used) when given a
function that is not itself.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>